refactor: repository pattern 도입
